Move ShopStack to Root, render ShopHome directly in tab
