+++ src/navigation/shop/ShopStack.tsx
@@ -3,16 +3,13 @@ import React from 'react';
 import {
   baseNavigatorOptions,
   baseScreenOptions,
-} from '../../../constants/NavigationOptions';
-import ShopHome, {ShopHomeParamList} from './ShopHome';
-import {HeaderTitle} from '../../../components/styled/Text';
+} from '../../constants/NavigationOptions';
+import {HeaderTitle} from '../../components/styled/Text';
 import {t} from 'i18next';
-import {NavigatorScreenParams} from '@react-navigation/native';
-import {CardConfigMap, GiftCard} from '../../../store/shop/shop.models';
-import ArchivedGiftCards from './gift-card/screens/ArchivedGiftCards';
+import {CardConfigMap, GiftCard} from '../../store/shop/shop.models';
+import ArchivedGiftCards from '../tabs/shop/gift-card/screens/ArchivedGiftCards';
 
 export type ShopStackParamList = {
-  Home: NavigatorScreenParams<ShopHomeParamList>;
   ArchivedGiftCards: {
     giftCards: GiftCard[];
     supportedGiftCardMap: CardConfigMap;
@@ -20,7 +17,6 @@ export type ShopStackParamList = {
 };
 
 export enum ShopScreens {
-  HOME = 'Home',
   ARCHIVED_GIFT_CARDS = 'ArchivedGiftCards',
 }
 
@@ -29,13 +25,12 @@ const Shop = createStackNavigator<ShopStackParamList>();
 const ShopStack = () => {
   return (
     <Shop.Navigator
-      initialRouteName={ShopScreens.HOME}
+      initialRouteName={ShopScreens.ARCHIVED_GIFT_CARDS}
       screenOptions={{
         ...baseNavigatorOptions,
         ...baseScreenOptions,
         gestureEnabled: false,
       }}>
-      <Shop.Screen name={ShopScreens.HOME} component={ShopHome} />
       <Shop.Screen
         name={ShopScreens.ARCHIVED_GIFT_CARDS}
         component={ArchivedGiftCards}
