+++ src/navigation/tabs/shop/ShopHome.tsx
@@ -28,7 +28,6 @@ import {
 import {APP_NETWORK} from '../../../constants/config';
 import {useAppSelector} from '../../../utils/hooks';
 import {StackScreenProps} from '@react-navigation/stack';
-import {ShopScreens, ShopStackParamList} from './ShopStack';
 import {useTranslation} from 'react-i18next';
 import {
   useFocusEffect,
@@ -40,6 +39,9 @@ import {useTheme} from 'styled-components';
 import {SlateDark, White} from '../../../styles/colors';
 import {sleep} from '../../../utils/helper-methods';
 import {Analytics} from '../../../store/analytics/analytics.effects';
+import {TabsScreens, TabsStackParamList} from '../TabsStack';
+
+export type ShopStackScreenProps = StackScreenProps<TabsStackParamList, TabsScreens.SHOP>;
 
 export enum ShopTabs {
   GIFT_CARDS = 'Gift Cards',
@@ -103,9 +105,7 @@ const getScrollViewHeight = (
     : getShopOnlineScrollViewHeight(integrationsCategories);
 };
 
-const ShopHome: React.FC<
-  StackScreenProps<ShopStackParamList, ShopScreens.HOME>
-> = ({route}) => {
+const ShopHomeScreen: React.FC<ShopStackScreenProps> = ({route}) => {
   const {t} = useTranslation();
   const theme = useTheme();
   const availableCardMap = useAppSelector(({SHOP}) => SHOP.availableCardMap);
@@ -322,4 +322,4 @@ const ShopHome: React.FC<
   );
 };
 
-export default ShopHome;
+export default ShopHomeScreen;
